requirements.txt
appnope==0.1.4
asttokens==3.0.1
bitarray==3.8.0
bitstring==4.3.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.1
comm==0.2.3
cryptography==46.0.4
debugpy==1.8.20
decorator==5.2.1
esptool==5.1.0
executing==2.2.1
idna==3.10
influxdb-client==1.50.0
intelhex==2.3.0
ipykernel==7.2.0
ipython==9.11.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.8.0
jupyter_core==5.9.1
markdown-it-py==4.0.0
matplotlib-inline==0.2.1
mdurl==0.1.2
nest-asyncio==1.6.0
numpy==2.3.4
packaging==26.0
pandas==3.0.1
parso==0.8.6
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.9.4
prompt_toolkit==3.0.52
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==3.0
Pygments==2.19.2
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-gitlab==3.15.0
PyYAML==6.0.3
pyzmq==27.1.0
reactivex==4.1.0
reedsolo==1.7.0
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.2
rich-click==1.9.7
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
stack-data==0.6.3
tornado==6.5.5
traitlets==5.14.3
typing_extensions==4.15.0
urllib3==2.5.0
wcwidth==0.6.0
websockets==15.0.1
